add navCtrl for show the active class in JS,
Change the nav display using AngularJS instead of original version

The active worked on the button without link, But not worked on the
button with link. Because the nav bar is generally for whole website,
not every single page, when jump to a new page, the nav bar will reset
to original version.